Rebasing the project on a new architecture
Changelog: I created a simple architecture where I'll base the System 11
kernel. This may change around in time and i want to be able to create a
dynamic yet simple and robust kernel to address "modern" features. Btw I
want to introduce a protected mode for the OS expecially for proventing
users to write in protected zone of the address space like the OS zone,
the zero page or the device address that need to be used by the kernel
and by each individual driver for them. To implement this I need an
external circuit and this is possible by now ONLY on 6502 machine with a
tweak using the nmi. By decoding the address visited by the cpu and by
setting and external register it's possible to apply a sort of
protection against the use of specific address spaces. But for now this
is just an idea and first I need to rebase the kernel and write it down
to a more robust and functional architecture.